% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfars.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Query accidents for a specific State and year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{The State ID number (from 1 to 56) for the State to be queried}

\item{years}{A four digit number of the year the accidents need to be extracted for}
}
\value{
This function prints a map of the specified State with dots for each documented accident
}
\description{
This function serves to extract fars accidents for a specific state and year and, if present,
   print a map with accidents
}
\note{
Error messages are displayed in case an invalid State id number is used as an argument or
   it's not present and in case no accidents are documented for the specified State and year
}
\examples{
\dontrun{fars_map_state(33, 2014)}

}
