% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfars.R
\name{fars_read}
\alias{fars_read}
\title{Check the existenence of a file and read it}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{The name of the file to be read. The filename should be built with the
"make_filename" function.}
}
\value{
This function returns a tibble (dataframe). If the filename provided as argument is wrong or the file
   is not present in current working directory, the function returns an error message.
}
\description{
This function takes as only argument a filename and first check if the inputted filename exists.
   If the file does not exist in current working directory an error message is given, otherwise the file is
   read as a dataframe
}
\examples{
\dontrun{data2013 <- fars_read("accident_2013.csv.bz2")}

}
