% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfars.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Iteratively read months and years from different filenames}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A list of years for which the observation needs to be extracted}
}
\value{
This function returns a list of data frames, one for each year of the argument list and
   containing the months and year for each observation. If one of the specified years is not present
   in the dataset an error message
}
\description{
This function is used to iteratively read files relative to fatalities in a series
of initially selected years, i.e. from different files referring to different years.
}
\examples{
\dontrun{multi_years <- fars_read_years(list(2013, 2014, 2015))}

}
