% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfars.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize accidents per year and month}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A list of years for the user wants to see the summary, or a single year expressed
as a four digit numeral}
}
\value{
This function returns a table with the months aggregates of accidents (in rows)
   for each year (in columns)
}
\description{
This function creates a summary table displaying the number of accidents for each month and year
}
\note{
Please note that if one of the specified years is not present in the dataset an error message
   is returned specifying which year is invalid (not found).
}
\examples{
\dontrun{summary1 <- fars_summarize_years(2014)}
\dontrun{summary2 <- fars_summarize_years(list(2013, 2014))}

}
