% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qfars.R
\name{make_filename}
\alias{make_filename}
\title{Build the filename for a specific year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The year, expressed in four digit, for which the filename string needs to be returned}
}
\value{
This function returns a character vector which is the filename for the specified year
}
\description{
This function takes as argument a four digit number indicating a year to build the complete
   filename for that year, according to FARS standard naming of datasets
}
\examples{
\dontrun{fars_read(2013)}

}
